+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Use `secondaryStrategy` when `recommendations` array is empty.
+
+
 ## [1.6.2] - 2022-01-10
 
 ### Fixed